Move function to module
